testing/web-platform/tests/orientation-event/orientation/window-property.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

function hasProperty(orientationProperty)
{
  for (let property in window) {
    if (property == orientationProperty)
      return true;
  }
  return false;
}

test(test => {
  assert_equals(typeof window.DeviceOrientationEvent, 'function');
  assert_false(window.propertyIsEnumerable('DeviceOrientationEvent'));
  assert_true('DeviceOrientationEvent' in window);
  assert_true(window.hasOwnProperty('DeviceOrientationEvent'));

  assert_equals(typeof window.ondeviceorientation, 'object');
  assert_equals(typeof window.ondeviceorientationabsolute, 'object');
  assert_true(hasProperty('ondeviceorientation'));
  assert_true(hasProperty('ondeviceorientationabsolute'));
  assert_true('ondeviceorientation' in window);
  assert_true('ondeviceorientationabsolute' in window);
  assert_true(window.hasOwnProperty('ondeviceorientation'));
  assert_true(window.hasOwnProperty('ondeviceorientationabsolute'));
}, 'Tests that the window.DeviceOrientationEvent and window.ondeviceorientation and window.ondeviceorientationabsolute properties are present.');
</script>
